Add demo index page
